hparams: extract Keras callback to separate module
Summary:
The `summary_v2` module no longer has any fundamental dependency on
TensorFlow. Its implementation still depends on `summary`, which does
directly reference the TensorFlow proto copies.

Part of a reorganization to match the structure described here:
<https://github.com/tensorflow/tensorboard/pull/2139#discussion_r279106791>

wchargin-branch: hparams-api-refactor
