+++ tensorboard/plugins/hparams/summary_v2_test.py
@@ -306,151 +306,5 @@ class DiscreteTest(test.TestCase):
       hp.Discrete(["one", 2])
 
 
-class KerasCallbackTest(tf.test.TestCase):
-
-  def setUp(self):
-    super(KerasCallbackTest, self).setUp()
-    self.logdir = os.path.join(self.get_temp_dir(), "logs")
-
-  def _initialize_model(self, writer):
-    HP_DENSE_NEURONS = hp.HParam("dense_neurons", hp.IntInterval(4, 16))
-    self.hparams = {
-        "optimizer": "adam",
-        HP_DENSE_NEURONS: 8,
-    }
-    self.model = tf.keras.models.Sequential([
-        tf.keras.layers.Dense(self.hparams[HP_DENSE_NEURONS], input_shape=(1,)),
-        tf.keras.layers.Dense(1, activation="sigmoid"),
-    ])
-    self.model.compile(loss="mse", optimizer=self.hparams["optimizer"])
-    self.callback = hp.KerasCallback(writer, self.hparams, group_name="psl27")
-
-  def test_eager(self):
-    def mock_time():
-      mock_time.time += 1
-      return mock_time.time
-    mock_time.time = 1556227801.875
-    initial_time = mock_time.time
-    with mock.patch("time.time", mock_time):
-      self._initialize_model(writer=self.logdir)
-      self.model.fit(x=[(1,)], y=[(2,)], callbacks=[self.callback])
-    final_time = mock_time.time
-
-    files = os.listdir(self.logdir)
-    self.assertEqual(len(files), 1, files)
-    events_file = os.path.join(self.logdir, files[0])
-    plugin_data = []
-    for event in tf.compat.v1.train.summary_iterator(events_file):
-      if event.WhichOneof("what") != "summary":
-        continue
-      self.assertEqual(len(event.summary.value), 1, event.summary.value)
-      value = event.summary.value[0]
-      self.assertEqual(
-          value.metadata.plugin_data.plugin_name,
-          metadata.PLUGIN_NAME,
-      )
-      plugin_data.append(value.metadata.plugin_data.content)
-
-    self.assertEqual(len(plugin_data), 2, plugin_data)
-    (start_plugin_data, end_plugin_data) = plugin_data
-    start_pb = metadata.parse_session_start_info_plugin_data(start_plugin_data)
-    end_pb = metadata.parse_session_end_info_plugin_data(end_plugin_data)
-
-    # We're not the only callers of `time.time`; Keras calls it
-    # internally an unspecified number of times, so we're not guaranteed
-    # to know the exact values. Instead, we perform relative checks...
-    self.assertGreater(start_pb.start_time_secs, initial_time)
-    self.assertLess(start_pb.start_time_secs, end_pb.end_time_secs)
-    self.assertLessEqual(start_pb.start_time_secs, final_time)
-    # ...and then stub out the times for proto equality checks below.
-    start_pb.start_time_secs = 1234.5
-    end_pb.end_time_secs = 6789.0
-
-    expected_start_pb = plugin_data_pb2.SessionStartInfo()
-    text_format.Merge(
-        """
-        start_time_secs: 1234.5
-        group_name: "psl27"
-        hparams {
-          key: "optimizer"
-          value {
-            string_value: "adam"
-          }
-        }
-        hparams {
-          key: "dense_neurons"
-          value {
-            number_value: 8.0
-          }
-        }
-        """,
-        expected_start_pb,
-    )
-    self.assertEqual(start_pb, expected_start_pb)
-
-    expected_end_pb = plugin_data_pb2.SessionEndInfo()
-    text_format.Merge(
-        """
-        end_time_secs: 6789.0
-        status: STATUS_SUCCESS
-        """,
-        expected_end_pb,
-    )
-    self.assertEqual(end_pb, expected_end_pb)
-
-  def test_explicit_writer(self):
-    writer = tf.compat.v2.summary.create_file_writer(
-        self.logdir,
-        filename_suffix=".magic",
-    )
-    self._initialize_model(writer=writer)
-    self.model.fit(x=[(1,)], y=[(2,)], callbacks=[self.callback])
-
-    files = os.listdir(self.logdir)
-    self.assertEqual(len(files), 1, files)
-    filename = files[0]
-    self.assertTrue(filename.endswith(".magic"), filename)
-    # We'll assume that the contents are correct, as in the case where
-    # the file writer was constructed implicitly.
-
-  def test_non_eager_failure(self):
-    with tf.compat.v1.Graph().as_default():
-      assert not tf.executing_eagerly()
-      self._initialize_model(writer=self.logdir)
-      with six.assertRaisesRegex(
-          self, RuntimeError, "only supported in TensorFlow eager mode"):
-        self.model.fit(x=[(1,)], y=[(2,)], callbacks=[self.callback])
-
-  def test_reuse_failure(self):
-    self._initialize_model(writer=self.logdir)
-    self.model.fit(x=[(1,)], y=[(2,)], callbacks=[self.callback])
-    with six.assertRaisesRegex(
-        self, RuntimeError, "cannot be reused across training sessions"):
-      self.model.fit(x=[(1,)], y=[(2,)], callbacks=[self.callback])
-
-  def test_invalid_writer(self):
-    with six.assertRaisesRegex(
-        self, TypeError, "writer must be a `SummaryWriter` or `str`, not None"):
-      hp.KerasCallback(writer=None, hparams={})
-
-  def test_duplicate_hparam_names_across_object_and_string(self):
-    hparams = {
-        "foo": 1,
-        hp.HParam("foo"): 1,
-    }
-    with six.assertRaisesRegex(
-        self, ValueError, "multiple values specified for hparam 'foo'"):
-      hp.KerasCallback(self.get_temp_dir(), hparams)
-
-  def test_duplicate_hparam_names_from_two_objects(self):
-    hparams = {
-        hp.HParam("foo"): 1,
-        hp.HParam("foo"): 1,
-    }
-    with six.assertRaisesRegex(
-        self, ValueError, "multiple values specified for hparam 'foo'"):
-      hp.KerasCallback(self.get_temp_dir(), hparams)
-
-
 if __name__ == "__main__":
   tf.test.main()
